backend/requirements.txt
Django>=3.0.8,<3.1.0
djangorestframework>=3.11.0,<3.20.0

djangorestframework-jwt>=1.11.0,<1.20.0
djangorestframework-simplejwt>=4.6.0,<4.7.0

djoser>=2.1.0,<2.2.0

django-cors-headers>=3.5.0,<3.6.0

psycopg2>=2.8.6,<2.9.0
psycopg2-binary>=2.8.6,<2.9.0

Pillow>=7.2.0,<7.3.0
uwsgi


dj-rest-auth>=2.1.1,<2.2.0

django-allauth>=0.44.0,<0.50.0

django-webpack-loader